+++ Deltos/Dialogs/ExportDialog.axaml.cs
@@ -64,6 +64,7 @@ public partial class ExportDialog: DialogWindow
         chTextFileTitle.IsChecked = fOptions.TextFileTitle.HasFlag(ExportTitleOptions.Title);
 
         chTreatTextFilesAsPlainText.IsChecked = fOptions.TreatTextFilesAsPlainText;
+        edtImageMaxWidth.Value = Math.Clamp(fOptions.ImageMaxWidth, 100, 2000);
     }
     /// <summary>
     /// Saves controls into options.
@@ -111,6 +112,7 @@ public partial class ExportDialog: DialogWindow
             fOptions.TextFileTitle |= ExportTitleOptions.Title;
 
         fOptions.TreatTextFilesAsPlainText = chTreatTextFilesAsPlainText.IsChecked == true;
+        fOptions.ImageMaxWidth = Math.Clamp((int)(edtImageMaxWidth.Value ?? 400), 100, 2000);
 
         if (fOptions.Language == ExportLanguage.None)
         {